additional translations
restore "App settings" key from `ui-agreements`, which feels wrong, but
is actually the smaller change since `settings.appSettings` is present
in that application but not here in `ui-licenses`, meaning it would be
necessary to add that key to all the locale files. Since licenses and
agreements tend to travel together, it seems like a low risk (indeed,
all this does is revert to the previous implementation).